Move shared utils to one file
It's easier to copy/paste one file in Azure Kudu.
